ci: set COMPOSE_COMMAND environment variable
Point the `ComposeHandler` to docker by setting the `COMPOSE_COMMAND`. Previously
it was attempted to use podman here by installing it via nix, however the GHA runner
configuration is a bit special and caused issues. As a result, we'll just use the
container tools that GH ships with their runners rather than trying to be fancy.

Signed-off-by: Ryan Drew <git@ry4n.me>